+++ arclet/alconna/builtin/commandline.py
@@ -3,9 +3,8 @@ from typing import List
 import re
 import json
 
-from arclet.alconna import Alconna, Args, Arpamar, Option, AnyUrl, AnyIP, AnyDigit, AnyFloat, AllParam, Email, Bool, \
-    command_manager, all_command_help, alconna_version
-from arclet.alconna.types import ArgPattern, PatternToken
+from arclet.alconna import Alconna, Args, Arpamar, Option, AllParam, command_manager, all_command_help, alconna_version
+from arclet.alconna.types import ArgPattern, PatternToken, AnyIP, AnyUrl, Email, Bool, AnyDigit, AnyFloat
 
 args_type = ArgPattern(
     r"(\[.+])*",