+++ dev_tools/test_alconna_2.py
@@ -1,7 +1,7 @@
 from typing import Union, Dict
 
 from arclet.alconna.builtin.construct import AlconnaString, AlconnaFormat
-from arclet.alconna.types import AnyStr, AnyDigit, UnionArg, pattern
+from arclet.alconna.types import UnionArg, pattern
 from arclet.alconna import Alconna, Args, Option
 from arclet.alconna import command_manager
 from graia.ariadne.message.element import At
@@ -15,7 +15,7 @@ ping1 = AlconnaString("ping <url:url>")
 Alconna.set_custom_types(digit=int)
 alc = AlconnaFormat(
     "lp user {target} perm set {perm} {default}",
-    {"target": AnyStr, "perm": AnyStr, "default": Args["de":bool:True]},
+    {"target": str, "perm": str, "default": Args["de":bool:True]},
 )
 alcc = AlconnaFormat(
     "lp1 user {target}",
@@ -28,7 +28,7 @@ debug(alc)
 alc.exception_in_time = False
 debug(alc.parse("lp user AAA perm set admin"))
 
-aaa = AlconnaFormat("a {num}", {"num": AnyDigit})
+aaa = AlconnaFormat("a {num}", {"num": int})
 r = aaa.parse("a 1")
 print(aaa)
 print(r)
@@ -54,17 +54,6 @@ def test_type(text: str):
 
 alc2 = Alconna("test", help_text="测试help直接发送") + Option("foo", Args["bar":str, "bar1":int:12345, "bar2":test_type])
 print(alc2.parse("test --help"))
-
-dic = alc1.to_dict()
-
-debug(dic)
-
-dic['command'] = 'test_type_1'
-
-alc3 = Alconna.from_dict(dic)
-print(alc3)
-print(alc3.get_help())
-print(alc3.parse("test_type_1 abcd 'testing a text' 2"))
 
 alc4 = Alconna(
     command="test_multi",